+++ src/vs/workbench/contrib/performance/electron-sandbox/performance.contribution.ts
