+++ src/vs/workbench/workbench.desktop.main.ts
@@ -140,7 +140,7 @@ import 'vs/workbench/contrib/externalTerminal/node/externalTerminal.contribution
 
 
 // Performance
-import 'vs/workbench/contrib/performance/electron-browser/performance.contribution';
+import 'vs/workbench/contrib/performance/electron-sandbox/performance.contribution';
 
 // CLI
 import 'vs/workbench/contrib/cli/node/cli.contribution';
